requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.6
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.2.0
packaging==24.2
phone_email_auth==0.13
phonenumbers==9.0.3
pillow==11.1.0
propcache==0.3.0
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sniffio==1.3.1
sqlparse==0.5.3
twilio==9.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.18.3
